+++ packages/cloudflare-pages/src/cloudflare-pages.ts
@@ -1,7 +1,7 @@
+import type { ConfigEnv, Plugin, ResolvedConfig, UserConfig } from 'vite'
 import { builtinModules } from 'module'
 import { readdir, writeFile } from 'node:fs/promises'
 import { resolve } from 'node:path'
-import type { ConfigEnv, Plugin, ResolvedConfig, UserConfig } from 'vite'
 import { getEntryContent } from './entry.js'
 
 type CloudflarePagesOptions = {
@@ -105,6 +105,7 @@ export const cloudflarePagesPlugin = (options?: CloudflarePagesOptions): Plugin 
         ssr: {
           external: options?.external ?? defaultOptions.external,
           noExternal: true,
+          target: 'webworker',
         },
         build: {
           outDir: options?.outputDir ?? defaultOptions.outputDir,