+++ packages/dev-server/e2e/e2e.test.ts
@@ -1,4 +1,4 @@
-import { test, expect } from '@playwright/test'
+import { test, expect, request } from '@playwright/test'
 
 test('Should return 200 response', async ({ page }) => {
   const response = await page.goto('/')
@@ -16,6 +16,22 @@ test('Should contain an injected script tag', async ({ page }) => {
 
   const lastScriptTag = await page.$('script:last-of-type')
   expect(lastScriptTag).not.toBeNull()
+
+  const nonce = await lastScriptTag?.getAttribute('nonce')
+  expect(nonce).toBeNull()
+
+  const content = await lastScriptTag?.textContent()
+  expect(content).toBe('import("/@vite/client")')
+})
+
+test('Should contain an injected script tag with a nonce', async ({ page }) => {
+  await page.goto('/with-nonce')
+
+  const lastScriptTag = await page.$('script:last-of-type')
+  expect(lastScriptTag).not.toBeNull()
+
+  const nonce = await lastScriptTag?.getAttribute('nonce')
+  expect(nonce).not.toBeNull()
 
   const content = await lastScriptTag?.textContent()
   expect(content).toBe('import("/@vite/client")')
@@ -65,22 +81,6 @@ test('Should return 200 response - /stream', async ({ page }) => {
   expect(content).toBe('Hello Vite!')
 })
 
-test('Should serve static files in `public/static`', async ({ page }) => {
-  const response = await page.goto('/static/hello.json')
-  expect(response?.status()).toBe(200)
-
-  const data = await response?.json()
-  expect(data['message']).toBe('Hello')
-})
-
-test('Should handle `env.ASSETS.fetch` function', async ({ page }) => {
-  const response = await page.goto('/assets/hello.json')
-  expect(response?.status()).toBe(200)
-
-  const data = await response?.json()
-  expect(data['message']).toBe('Hello')
-})
-
 test('Should return a vite error page - /invalid-response', async ({ page }) => {
   const response = await page.goto('/invalid-response')
   expect(response?.status()).toBe(500)
@@ -93,57 +93,6 @@ test('Should return a vite error page with stack trace - /invalid-error-response
   const response = await page.goto('/invalid-error-response')
   expect(response?.status()).toBe(500)
   expect(await response?.text()).toContain('e2e/mock/worker.ts')
-})
-
-test('Should set bindings from wrangler.toml [vars]', async ({ page }) => {
-  const res = await page.goto('/env', { waitUntil: 'domcontentloaded' })
-  expect(res?.ok()).toBe(true)
-  const json = await res?.json()
-  expect(json).toBeTruthy()
-  expect(json.env).toHaveProperty(
-    'VARIABLE_FROM_WRANGLER_TOML',
-    'VARIABLE_FROM_WRANGLER_TOML_VALUE'
-  )
-})
-
-test('Should set bindings from wrangler.toml [[d1_database]]', async ({ page }) => {
-  const res = await page.goto('/env', { waitUntil: 'domcontentloaded' })
-  expect(res?.ok()).toBe(true)
-  const json = await res?.json()
-  expect(json).toBeTruthy()
-  expect(json.env).toHaveProperty('DB_FROM_WRANGLER_TOML')
-})
-
-test('Should set bindings from root `env` in config', async ({ page }) => {
-  const res = await page.goto('/env', { waitUntil: 'domcontentloaded' })
-  expect(res?.ok()).toBe(true)
-  const json = await res?.json()
-  expect(json).toBeTruthy()
-  expect(json.env).toHaveProperty('ENV_FROM_ROOT', 'ENV_FROM_ROOT_VALUE')
-})
-
-test('Should set bindings from `cf` in config', async ({ page }) => {
-  const res = await page.goto('/env', { waitUntil: 'domcontentloaded' })
-  expect(res?.ok()).toBe(true)
-  const json = await res?.json()
-  expect(json).toBeTruthy()
-  expect(json.env).toHaveProperty('ENV_FROM_DEPRACATED_CF', 'ENV_FROM_DEPRACATED_CF_VALUE')
-})
-
-test('Should set bindings from `plugins` in config', async ({ page }) => {
-  const res = await page.goto('/env', { waitUntil: 'domcontentloaded' })
-  expect(res?.ok()).toBe(true)
-  const json = await res?.json()
-  expect(json).toBeTruthy()
-  expect(json.env).toHaveProperty('ENV_FROM_PLUGIN', 'ENV_FROM_PLUGIN_VALUE')
-})
-
-test('Should set bindings from `plugins` in config (async)', async ({ page }) => {
-  const res = await page.goto('/env', { waitUntil: 'domcontentloaded' })
-  expect(res?.ok()).toBe(true)
-  const json = await res?.json()
-  expect(json).toBeTruthy()
-  expect(json.env).toHaveProperty('ENV_FROM_PLUGIN_AS_FUNC', 'ENV_FROM_PLUGIN_AS_FUNC_VALUE')
 })
 
 test('Should set `workerd` as a runtime key', async ({ page }) => {
@@ -161,4 +110,28 @@ test('Should not throw an error if accessing the `caches`', async ({ page }) => 
   // Cache API provided by `getPlatformProxy` currently do nothing.
   // It does **not** return cached content.
   expect(await resCached?.text()).not.toBe('cached')
+})
+
+test('Should set `cf` properties', async ({ page }) => {
+  const res = await page.goto('/cf')
+  expect(res?.ok()).toBe(true)
+  expect(await res?.json()).toEqual({ cf: true })
+})
+
+test('Should return files in the public directory', async ({ page }) => {
+  const res = await page.goto('/hono-logo.png')
+  expect(res?.status()).toBe(200)
+})
+
+test('Should not crash when receiving a HEAD request', async () => {
+  const apiContext = await request.newContext()
+  const response = await apiContext.fetch('/', {
+    method: 'HEAD',
+  })
+
+  expect(response.status()).toBe(200)
+  expect(response.headers()['content-type']).toMatch(/^text\/html/)
+  const body = await response.body()
+  expect(body.length).toBe(0)
+  await apiContext.dispose()
 })