+++ packages/dev-server/src/cloudflare-pages/index.ts
@@ -1,3 +1,0 @@
-export { getEnv } from './cloudflare-pages.js'
-import plugin from './cloudflare-pages.js'
-export default plugin