+++ packages/dev-server/src/types.ts
@@ -1,3 +1,5 @@
+import type { ViteDevServer } from 'vite'
+
 export type Env = Record<string, unknown> | Promise<Record<string, unknown>>
 export type EnvFunc = () => Env | Promise<Env>
 export type GetEnv<Options> = (options: Options) => EnvFunc
@@ -12,10 +14,7 @@ export type Fetch = (
   executionContext: ExecutionContext
 ) => Promise<Response>
 
-export interface Plugin {
-  env?: Env | EnvFunc
-  onServerClose?: () => void | Promise<void>
-}
+export type LoadModule = (server: ViteDevServer, entry: string) => Promise<{ fetch: Fetch }>
 
 export interface Adapter {
   /**